add basic abstract command
