+++ src/App/AbstractCommands/AbstractDeleteCommand.php
@@ -2,10 +2,7 @@
 
 namespace Lio\App\AbstractCommands;
 
-use Exception;
-use GuzzleHttp\Exception\BadResponseException;
 use GuzzleHttp\Exception\GuzzleException;
-use Lio\App\Console\Command;
 use Psr\Http\Message\ResponseInterface;
 use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Input\InputInterface;
@@ -13,12 +10,8 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 
-abstract class AbstractDeleteCommand extends Command
+abstract class AbstractDeleteCommand extends AbstractCommand
 {
-	/**
-	 * @var string
-	 */
-	protected $apiEndpoint = '';
 
 	protected function configure()
 	{
@@ -28,35 +21,18 @@ abstract class AbstractDeleteCommand extends Command
 
 	/**
 	 * @param InputInterface $input
-	 * @param OutputInterface $output
-	 * @return int|void|null
+	 * @return ResponseInterface
 	 * @throws GuzzleException
-	 * @throws Exception
 	 */
-	protected function execute(InputInterface $input, OutputInterface $output)
+	protected function sendRequest(InputInterface $input): ResponseInterface
 	{
-		parent::execute($input, $output);
-		if (!$this->askConfirm('<info>Are you sure you want to delete ? (y/N)</info>', $output, $input)) {
-			return 0;
-		}
-		try {
-			$response = $this->httpHelper->getClient()->request(
-				'DELETE',
-				$this->getApiEndpoint(),
-				[
-					'headers' => $this->httpHelper->getHeaders(),
-				]
-			);
-			if (!empty($input->getOption('json'))) {
-
-				$output->writeln($response->getBody()->getContents());
-			} else {
-				$this->renderOutput($response, $output, $input);
-			}
-		} catch (BadResponseException $badResponseException) {
-			$output->writeln('<error>' . $badResponseException->getResponse()->getBody()->getContents() . '</error>');
-			return 1;
-		}
+		return $this->httpHelper->getClient()->request(
+			'DELETE',
+			$this->getApiEndpoint(),
+			[
+				'headers' => $this->httpHelper->getHeaders(),
+			]
+		);
 	}
 
 	/**
@@ -75,29 +51,4 @@ abstract class AbstractDeleteCommand extends Command
 		$question = new ConfirmationQuestion($questionText, false);
 		return $helper->ask($input, $output, $question);
 	}
-
-	/**
-	 * @param string $apiEndpoint
-	 */
-	public function setApiEndpoint(string $apiEndpoint): void
-	{
-		$this->apiEndpoint = $apiEndpoint;
-	}
-
-	/**
-	 * @return string
-	 */
-	public function getApiEndpoint(): string
-	{
-		return $this->apiEndpoint;
-	}
-
-
-	/**
-	 * @param ResponseInterface $response
-	 * @param OutputInterface $output
-	 * @param InputInterface $input
-	 * @return null
-	 */
-	abstract protected function renderOutput(ResponseInterface $response, OutputInterface $output, InputInterface $input);
 }
